requirements.txt
Flask>=2.1
langchain==0.3.21
pyttsx3==2.90
google-generativeai==0.0.6  
langchain-google-genai==0.0.6  
python-dotenv==0.19.1
